refactor(core): extract Blockchain MDBX storage I/O to own module
Five impl Blockchain methods that touch the MDBX storage handle move
to crates/sentrix-core/src/blockchain_storage_io.rs:

  init_storage_handle  — bind Arc<MdbxStorage> at boot
  persist_block_durable — atomic block bytes + hash idx + height + sync
  record_tx_in_index   — fire-and-warn tx->height write
  lookup_tx_in_storage — read-side fallback for evicted blocks
  backfill_txid_index  — one-shot startup walk with warm-chain fast path

Rust permits a single impl Blockchain {} span multiple modules within
the same crate, so all bc.method(...) call sites keep resolving
unchanged. blockchain.rs drops from 3406 -> 3201 LOC (-6%); 215 LOC
moved to the new sibling module. height_key / key_to_height imports
moved with them (they're no longer used in blockchain.rs).

231 / 231 sentrix-core tests still green; no behaviour change.
